fix: 🐛 Enable testing localized code (#44)
* fix: 🐛 Enable testing localized code

Currently it's not possible to test code using `$Dictionary` (e.g.
components with calls to `this.localize()`), as the dictionary is
generated during build by `gulp-tasks`. The new function
`_generateDictionary` generates this dictionary during init.

* refactor: 💡 Move code to its own file

Move `generateDictionary()` to its own file to make it testable.

* test: 💍 Test `generateDictionary()`

* docs: ✏️ Add new config value to README

* fix: 🐛 Don't fail on undefined languages

* test: 💍 Improve tests

Add a check of language values to app test.